// local-hobbies-frontend/src/types/user.ts
export enum UserGenderMatched {
    Man = 'Man',
    Nonbinary = 'Nonbinary',
    Woman = 'Woman',
}
// ...
export interface UserOnboardingRequest {
    name?: string;
    /** ISO 8601 date (YYYY-MM-DD) */
    birthDate?: string;
    location?: GeoJsonPoint;
    publicContactInfo?: string;
    genderMatched?: UserGenderMatched;
    showAge?: boolean;
    showGenderDisplayed?: boolean;
    hobbies?: HobbyCreationRequest[];
    availabilities?: AvailabilityOnboardingRequests;
}

/**
 * User types
 */

export interface UserUpdateRequest {
    email?: string;
    name?: string;
    /** ISO 8601 date */
    birthDate?: string;
    genderDisplayed?: string | null;
    bio?: string | null;
    location?: GeoJsonPoint;
    publicContactInfo?: string;
    /** Key for cloud storage, null if removing */
    profilePhotoKey?: string | null;
    genderMatched?: UserGenderMatched;
    showAge?: boolean;
    showGenderDisplayed?: boolean;
}